+++ partiql-ast/tests/test_ast.rs
@@ -1,8 +1,6 @@
 mod common;
 
-use partiql_ast::ast;
 use partiql_ast::ast::*;
-use partiql_source_map::location::{BytePosition, Location};
 
 #[test]
 fn test_ast_init() {
@@ -10,33 +8,18 @@ fn test_ast_init() {
 
     let _i = Item {
         kind: ItemKind::Query(Query {
-            set: QuerySet::Expr(Box::new(Expr {
-                kind: ExprKind::Lit(Lit::Int32Lit(23).to_ast(BytePosition::from(1)..12.into())),
-            }))
-            .to_ast(BytePosition::from(1)..12.into()),
+            set: AstNode {
+                id: NodeId(2),
+                node: QuerySet::Expr(Box::new(Expr {
+                    kind: ExprKind::Lit(AstNode {
+                        id: NodeId(1),
+                        node: Lit::Int32Lit(23),
+                    }),
+                })),
+            },
             offset: None,
             order_by: None,
             limit: None,
         }),
     };
-
-    let span_only = ast::SymbolPrimitive {
-        value: "symbol1".to_string(),
-        case: Some(CaseSensitivity::CaseInsensitive),
-    }
-    .to_node()
-    .location(Location {
-        start: BytePosition::from(12),
-        end: BytePosition::from(1),
-    })
-    .build()
-    .expect("Could not retrieve ast node");
-
-    assert_eq!(
-        Some(Location {
-            start: BytePosition::from(12),
-            end: BytePosition::from(1),
-        }),
-        span_only.location
-    );
 }