+++ partiql-source-map/src/lib.rs
@@ -2,3 +2,4 @@
 
 pub mod line_offset_tracker;
 pub mod location;
+pub mod metadata;