+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- *BREAKING:* Refactors the AST
+  - Removed Location from the AST, replacing with a 'node id' that gives the AST node identity; the id can be used to retrieve Location
+
 ### Added
 - Adds the following functionalities to PartiQL Playground:
   - Moves the project to a Node.js project